# routers/binance_data.py
from fastapi import APIRouter, Depends, HTTPException
from typing import List, Optional
from datetime import datetime
from schemas.binance_data import MarketData, HistoricalDataResponse, Kline

router = APIRouter()

@router.get("/market-data/real-time", response_model=MarketData)
async def get_real_time_market_data(
    symbol: str,
    interval: Optional[str] = "1m"
):
    """
    Fetch real-time market data for a given symbol.

    Parameters:
    - symbol: Trading pair symbol (e.g., 'BTCUSDT')
    - interval: Time interval for the data (default: '1m')

    Returns:
    - dict: Real-time market data including price and volume
    """
    try:
        # TODO: Implement Binance API integration
        return {
            "symbol": symbol,
            "interval": interval,
            "timestamp": datetime.now().isoformat(),
            "price": 0.0,  # Placeholder
            "volume": 0.0  # Placeholder
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/market-data/historical", response_model=HistoricalDataResponse)
async def get_historical_market_data(
    symbol: str,
    start_time: datetime,
    end_time: datetime,
    interval: Optional[str] = "1h"
):
    """
    Fetch historical market data for a given symbol and time range.

    Parameters:
    - symbol: Trading pair symbol (e.g., 'BTCUSDT')
    - start_time: Start time for historical data
    - end_time: End time for historical data
    - interval: Time interval for the data (default: '1h')

    Returns:
    - dict: Historical market data including OHLCV data
    """
    try:
        # TODO: Implement historical data fetching
        return {
            "symbol": symbol,
            "interval": interval,
            "start_time": start_time.isoformat(),
            "end_time": end_time.isoformat(),
            "data": []  # Placeholder for historical data
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))