# routers/trading.py
from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.orm import Session
from typing import List

from database import get_db
from models.trading_crew import TradingCrew
from schemas.trading import TradingCrewCreate, TradingCrewResponse

router = APIRouter()

@router.post("/crews", response_model=TradingCrewResponse)
async def create_trading_crew(
    crew: TradingCrewCreate,
    db: Session = Depends(get_db)
):
    """Create a new trading crew"""
    pass  # TODO: Implement creation logic

@router.get("/crews", response_model=List[TradingCrewResponse])
async def get_trading_crews(
    db: Session = Depends(get_db)
):
    """Get all trading crews"""
    pass  # TODO: Implement retrieval logic

@router.get("/crews/{crew_id}", response_model=TradingCrewResponse)
async def get_trading_crew(
    crew_id: int,
    db: Session = Depends(get_db)
):
    """Get a specific trading crew"""
    pass  # TODO: Implement retrieval logic

@router.put("/crews/{crew_id}/activate")
async def activate_trading_crew(
    crew_id: int,
    db: Session = Depends(get_db)
):
    """Activate a trading crew"""
    pass  # TODO: Implement activation logic

@router.put("/crews/{crew_id}/deactivate")
async def deactivate_trading_crew(
    crew_id: int,
    db: Session = Depends(get_db)
):
    """Deactivate a trading crew"""
    pass  # TODO: Implement deactivation logic